补全数据库账户与密码脚本，增加将数据库配置文件同步到其他服务器选项

+++ scripts/os_tool/vm/bullseye/repair__fix_mysql_apt_db_user.sh
@@ -9,12 +9,17 @@
     python3 /www/server/jh-panel/plugins/mysql-apt/index.py fix_all_db_user
     popd > /dev/null
 
+
+
+    
+    filename="/www/server/mysql-apt/mysql.db"
     # 提示是否rsync同步
-    read -p "需要将数据库信息文件/www/server/mysql-apt/mysql.db同步到其他服务器吗?(默认n)[y/n]" sync
+    read -p "需要将数据库信息文件${filename}同步到其他服务器吗?(默认n)[y/n]" sync
     sync=${sync:-n}
 
 
     if [ $sync == "y" ]; then
+        sync_file=""
         # 输入目标服务器IP
         read -p "请输入备用服务器IP: " remote_ip
         if [ -z "$remote_ip" ]; then
@@ -26,6 +31,10 @@
         read -p "请输入备用服务器SSH端口(默认: 10022): " remote_port
         remote_port=${remote_port:-10022}
 
-        rsync -avu -e "ssh -p ${remote_port}" --progress /www/server/mysql-apt/mysql.db root@${remote_ip}:/www/server/mysql-apt/mysql.db
+        read -p "确定要同步${filename}到${remote_ip}:${filename}吗?(默认n)[y/n]" confirm
+        confirm=${confirm:-n}
+        if [ $confirm == "y" ]; then
+            rsync -avu -e "ssh -p ${remote_port}" --progress /www/server/mysql-apt/mysql.db root@${remote_ip}:/www/server/mysql-apt/mysql.db
+        fi
     fi
 # fi
